102.2 Creating an Image Card ComponentIssues with Grid CSS(Section 10: Using Ref's for DOM Access)
Let's continue working on sync because we've got a couple more things to do now.

The first thing I know without doubt I can tell you right now we're going to have to refer to this image

objects several times inside this render method rather than referring back to it with this stop Propp

site image over and over let's deep structure out just the properties we care about.

Off of this stop Propst image as we did before.

Back over here inside of our map function so inside my render method Osei Konst description and your

els.

And that's going to come from this stock prop's dot image.

And now inside of this tag right here I can reduce it down to just description and down to just your

Elle's regular OK I'll save.

And it looks like we condensed to that right there.

OK.

So looks pretty good.

Let's take a quick pause.

When we come back the next section we're going to add a little bit more code to our image card to make

it look just a little bit better and we'll start to focus on adding in all of that grid Arturo's and

the grid row and did all that good stuff.

So quick pause and I'll see you in just a minute.