106.1 Dynamic Spans(Section 10: Using Ref's for DOM Access)
In the last section we got the real height of each image after it was actually loaded into the browser.

So now we're going to use that height to figure out how many spans each of these images should occupy.

As a quick reminder let me walk you through that process really quickly to say you recall why this is

relevant for the grid CSSA stuff.

Zongo and to expect or inspect this div right here.

All right so this div is inside of the image list.

Remember we had set on the image list class.

The thing that wraps all these different images.

We had set a grid auto rows of 150 pixels.

That means that every single one of these little tiles here or each of these cells is going to be only

150 pixels large.

Now that's going to cause some these images to be cut off.

So in order to get around that we can set a grid row and property with a value of like span two and

that's going to allocate two cells worth for that particular image.

If we did span three it would give us even a little bit more.

So now that we have the height of this image and we know exactly how tall each row is because we set

that as a rule inside of our CSSA file right here we know each row is equal to 150 pixels.

We can use the image height and the size of each row or the height of each row and essentially just

figure out how many rows it takes to display each of these images.

That's it.

OK.

So to do so back inside of image card here's our set span's.

We're going to do a let's first pull off just the height here really quickly and assign it to a variable

like so.

And then to get the number of spans that this thing requires.

I'll do my height divided by the row height which is 150 pixels.

And then I'm going to add one to this.

The reason that I'm going to add one is to make sure that if we have like a portion of a row that this

image needs it's going to be rounded up or essentially a ghost like the next highest row and then just

to make sure I'm going to do a math dot ceiling on this thing just to cap its value and you it actually

right now let's remove that plus one just so we can see what happens when we just start with this basic

instrumentation right here.