106.2 Dynamic Spans(Section 10: Using Ref's for DOM Access)
OK so after we calculate the spans we'll then set it on our states X-O notice how we have a identical

key and value here so we can use a little bit of E.S. 2015 syntax to shorten it just be span's like

so now and those I've used state without initializing state inside of my constructor first.

So I can very quickly take care of that inside my constructor.

I'll say this Scott States is span's and I'll defer all its valid to be zero.

Like so.

So until we actually load the image and determine how much height it requires We're going to say this

thing basically requires zero space which is totally fine.

So the last thing we do.

Is take that spans value and assign it as a style to the div right here.

So essentially remember right here I just mentally put in grid row and span three.

You want to take that number of spans and essentially insert it right there span three or four or five

or whatever it is.

So to do so on this div I'm going to pass in a inline style.

It's going to have grid row ends.

And then for the value will you say yes 2015 template string.

So I use a back tick right there.

Remember a back tick is the character to the left of the one on your keyboard.

It is not a single quote.

It is a back tick.

And then I'll say span.

And then just the number of spans this thing requires so dollar sign curly braces this state that spans.

All right.

That's it.

So let's save this.

We might need a little bit of tweaking in here but that's totally fine.

Let's just make sure that this works as it stands.

OK.

So I can flip back over it for Kerr's again.

All right.

Now you'll notice that in some cases there's a little too much space.

But in other cases it's just right.

So it looks like we're really close to the last thing we're going to do here is going to be just a little

bit of a trick.

OK.

That sounds to me a little bit of a trick.

It is kind of interesting.

So remember when we set on the image list div like the thing that wraps all these images we had said

grid auto rows 150 pixels but we actually chose that kind of arbitrarily and now with the 150 pixels.

In some cases we are getting too much white space underneath it.