106.3 Dynamic Spans(Section 10: Using Ref's for DOM Access)
So one way that we could fix this would be to make these rows actually much smaller like 10 pixels.

If we only use rows of 10 pixels that means that we're going to get a lot of fine grained detail and

make sure that we do not somehow over allocate too many rows to any individual image.

Let's make these changes and you'll very quickly see what I mean.

So back inside my image list at CSSA file I'm going to find grid auto rows which allocates space to

each row.

I'm going to change it down to be just 10 pixels and then back inside of my image cart.

When we take our height and divide by 150 remember that 150 right there was to figure out how many rows

this image needed to be displayed properly.

So now we don't have rows of 150 pixels.

We have rows of 10 pixels.

So going to change that to 10 as well.

The net result is that each image is going to take up far more spans than it did before which is totally

fine.

So now save this will flip back over.

Try it again.

And while it still looks like it's a little bit off.

But don't worry.

This is very easily fixed if you do another kind of mouse over here and look at that component.

Remember that we had set up that grid gap rule right here to allocate 10 pixels between each one of

these cells.

So now that we have rows of only 10 pixels the Greek gap of 10 pixels is essentially doubling the amount

of whitespace we need.

So the last thing we're going to do to fix this is make sure that in the vertical direction or between

rows we do not have any grid gap.