108. App Overview 109. Component Design(Section 11: Let's Test Your React Mastery! )
In the last section we finished up our last application in this video we're going to start to move on

to our next app.

Now quick note here at the very start of this section I want you to know that this is an optional application.

This is an app that I put together very quickly just so you can get some of a better idea or some practice

on some base low level reac concepts.

Everything that we've gone through the course so far in practice this application is going to be very

very similar to the one that we just put together with just a couple of different design patterns being

used in the different source of data that we're going to pull from instead of the splash API.

So if you feel as though you already understand all the basics every act like classes state props inputs

all that kind of good stuff.

Fantastic.

You can always pause the video right now and move on to the next section and take a look at the next

app we're going to build.

OK.

So if you want to skip go ahead and do it right now.

Otherwise we're going to talk about what we're going to making inside this video.

All right so here we go here's our Mocca.

This is essentially a YouTube browser application in code and approach and design.

It's going to be somewhat similar to the on Splash app that we put together just a little bit ago at

the very top a user is going to enter in some search term such as in this case I put in buildings for

this example.

They're then going to hit the enter key which is going to trigger a search request rather than the splash

API.

We're going to make a request over to the YouTube public free API with the YouTube API.

We're going to do a search for a list of videos matching this search term right here.

Once we get a list of videos back from this YouTube API we're going to show them on the right hand side

of the screen.

A user can then click on one of these videos and we will feature it right in the center so the user

will see the video appear right here.

They can click on a play button and play the video will also show them a title and a description of

the video at the bottom as well.

So as you see here some similar elements to what we were doing.

We're still going to have a search bar at the top.

We're going to make a request to an outside API.

We're going to render a list of items and then from there we're going to learn a couple new small things

around clicking on an item and then making sure that we show a video on the screen.

So it is similar.

Just one or two little additions to this application.

Like I said this is mostly an application so that you can just solidify some of the fundamental knowledge

around it.

All right.

So that's it.

Let's take a quick pause right here.

When we come back the next video we'll start talking about the design approach around this application

and also start to generate our new app so quick pause and I'll see you in just a minute.

109. Component Design

In the last video we started talking about the next application we're going to work on in this section

we're going to talk about the design and approach and we'll go ahead and generate our app.

All right so we're going to take this mock up right here and break it down into a couple of separate

components so that we get a good idea of what we're going to be building as we work on the application.

All right.

So here's my idea on how we're going to split this up into separate components.

I think that we're still going to have a search bar component display at the very top of the screen

and the search bar is going to be essentially identical in functionality to what we had in the last

application.

It's going to show some text input.

A user can type in there that's going to update our state and then any time the user hits the enter

key enter key that will trigger a form submit event and send that search term back up to some other

app component or parent component which will then take that search term and do an actual API request.

Now on the right hand side we're going to have a video list that's going to render the entire list of

videos over here.

And then each of those individual parts or each of those individual boxes we will refer to as a video

item.

So if video item is one particular video all the video items are going to be contained by the video

list.

So just to be clear I only drew the screen box around here or right here around this last little item.

But this right here would also be a video item and a video item right there to pay.

And then finally we're going to add the video detail which will be responsible for actually showing

a video player that can play the actual YouTube video that has been fetched and the title description

right underneath that now.

So we're really clear here these are going to be the real videos that we are playing from YouTube.

It's not any fake video or something like that.

We need to actually figure out how we're going to show a video directly from YouTube and embed it essentially

inside of our application.

But that's going to be way more easy and straightforward than you might imagine.

All right now the very last thing I want to show you is our component hierarchy.

So at the top are still going to have that over Archaean app component which is responsible for holding

all the state of our application for retrieving a list of videos and first storing whatever the current

search term is.

The app is then going to configure the search bar video detail and video lists by passing props from

the app down to each of those.

And in turn the video list will pass some props down to the video item or the collection of video items

so we have in order to properly display each of those separate videos.

All right.

That's it.

Let's take a pause right here.

Actually one last thing.

Let's make sure we generate our application first.

It's going to flip on over to the terminal and go to some still inside of my pics project directory

and go up one folder back to my workspace directory and then once I'm back up your itinerary a new project

with create re-act app and we'll call this one.

Videos like so.

All right.

So that's going to generate a project.

We'll take a quick pause when we come back the next video.

We're going to start putting this thing together.